ecs/requirements.txt
boto3
ffmpeg-python
openai-whisper
numpy
